Add the '--no-merge' and '--bridge-file' options.
Add two more options to the 'sssom-inject' command:

* '--no-merge' to avoid merging the SSSOM-derived bridging axioms into
  the current ontology;
* '--bridge-file' to write the bridging axioms into a separate ontology.

This is to provide more flexibility in how the sssom-inject may be used.
The default is to inject the bridging axioms into the current ontology.
By adding '--bridge-file', the same command can in addition be used to
generate a small bridge file similar to the ones currently produced in
Uberon. By adding '--no-merge' on top of that, the command will *only*
generate the extra bridge file, without modifying the current ontology.