Generate IAO_0000589 annotation when creating bridges.
When creating a cross-species bridge, generate an annotation with
IAO_0000589 property ("OBO unique label"), if a taxon name was provided
(a taxon name may be provided after the taxon IRI in the
`--cross-species` option, as in `--cross-species
NCBITaxon:7227,Drosophila`).

Remove the IAxiomGenerator interface. The way bridging axioms and
bridges are generated is still very much work in progress, it's too
early to fix the API in interfaces.