Add the IMappingFilter and IMappingTransformer interfaces.
Add an interface to represent an arbitrary filter of mappings (take
mapping and return true or false), and an interface to represent a
generic transformer of mappings (take a mapping and generate another
object from it).

These interfaces are intended to be the building blocks for all
operations on mappings.

Also add two implementations of IMappingTransformer: one that generate
OWL equivalent axioms and one that generate IAO_0000589 annotation
assertion axioms.