Add the OWLGenerator class.
We add a class to encapsulate the entire OWL axiom generation process as
a rule-applying engine. Roughly, the generator is provided with a list
of rules that determines which axioms to generate for which mappings.

The rules themselves are represented by the new generic class
MappingProcessingRule<T>. Strictly speaking, for now this class does not
need to be exposed (it could stay hidden as an implementation detail of
the OWLGenerator), but a similar logic may later be used to generate
something else than OWL axioms from a set of mappings.