Handle predicate_modifier when filtering.
When filtering on predicate, if the predicate matches the pattern, check
whether the predicate is negated (predicate_modifier=Not); if if it,
then the filter should reject the mapping.

Considering a mapping where the predicate is 'skos:exactMatch' with a
predicate_modifier set to 'Not':

* a filter 'predicate==skos:exactMatch' will REJECT the mapping;

* a filter '!predicate==skos:exactMatch' will ACCEPT the mapping (the
  predicate filter will initially reject it because of the previous
  rule, then the negated filter will invert the result).

This means that '!predicate==skos:exactMatch' will not only select all
the mappings that have a predicate other than 'skos:exactMatch', but
also the mappings that have a negated 'skos:exactMatch' predicate.

To select only the mappings that really have a predicate other than
'skos:exactMatch' ("positive" non-skos:exactMatch mappings), one must
use '(!predicate==skos:exactMatch) && (!predicate_modifier==Not').