Expand CURIE(s) within the OWL expression.
When parsing a SSSOM/T-provided Manchester expression, expand any CURIE
that may be found inside the expression.

This is so that OWL expressions in a SSSOM/T ruleset may make use of
such CURIEs to avoid using full-length identifiers. That is, it is now
possible to write this:

  gen("%subject EquivalentTo %object and
       (BFO:0000050 some NCBITaxon:7227");

instead of that:

  gen("%subject EquivalentTo %object and
       (<http://purl.obolibrary.org/obo/BFO_0000050> some
        <http://purl.obolibrary.org/obo/NCBITaxon:7227>)");