Try recognising common Manchester patterns.
The way the default OWL axiom generator works is pretty inefficient: we
need to invoke a full-blown Manchester syntax parser every time we have
an axiom to generate on a mapping -- we cannot parse the expression once
and for all because it has to be parsed *after* the %subject and %object
placeholders have been substituted).

So we'll try avoiding using the default generator when possible, by
having a quick, regex-based look on the expression first to see if it
can be handled by a more efficient generator.

For now, we only recognise annotation assertion expressions (%subject
Annotation REL:1234 "value"), which cannot be handled by the default
generator anyway.