Change the way we deal with undeclared prefix names.
When the prefix manager encounters an undeclared prefix name, do not
throw an exception immediately. Instead, record the undeclared prefix
name expose the list of all undeclared prefix names to client code.

Then, have both the SSSOM TSV reader and the SSSOM/Transform reader
check that list once they are done parsing, and report the error
accordingly (by throwing a SSSOMFormatException for the SSSOM TSV
reader, adding it to the list of parsing errors for the SSSOM/Transform
reader).

This allows to report to the user *all* undeclared prefix names, instead
of erroring out on the first one that is encountered.