Add the --dispatch-table option.
Allows to automatically dispatch the axioms generated from mappings into
an arbitrary number of files based on the tags associated to the SSSOM/T
rules.

When a dispatch table is used, and an axiom is generated by a rule
carrying one or several tag(s), the axiom will be written to the
ontology associated with each tag, if such an ontology is specified in
the dispatch table.