Implement "direct" mapping translation.
Add a new axiom generator that produces axioms based on the "OWL
reification" rules set forth in the SSSOM standard.

Add a `--direct` option to the sssom-inject command to apply that
generator to all mappings in a set.

Recognise the 'direct' keyword, when used on its own, in the `gen`
command to set the direct generator as the generator of a SSSOM/T rule.