Add preliminary support for annotating axioms.
When generating axioms according to the "direct" rules from the SSSOM
specification, start annotating them with the mapping metadata, as
mandated by the specification.

This is very preliminary as we currently only annotate using the
mapping_justification, author_id, and object_label metadata. Adding all
the metadata will be a tedious task, but I am not sure yet how much of
it can be factored.