Allow arbitrary function names in SSSOM/Transform.
Major change to the SSSOM/Transform language. Instead of allowing only
the pre-defined function `gen` (in addition to the preprocessing
instructions `stop` and `invert`), we now allow arbitrary function
names.
The SSSOMTransformReader class is made abstract, so that client code has
to extend it to provide the code required to handle client-defined
functions.