Do not store quoted IRIs in the entity checker.
The Manchester parser queries the entity checker with quoted IRIs if
that's how it found them in the source text. We previously stored entity
IRIs in quoted form in the entity checker, so that we could directly
compare with the value provided by the parser. But then we used that
same value to create the IRI needed by the OWL Data Factory, resulting
in IRIs of the form '<<http...>>' (i.e., with a pair of quote being part
of the IRI).

To avoid that, we rather store naked IRIs in the entity checker, and
take care of removing the quotes from the value we get from the parser
before checking if that value corresponds to a known IRI.