Create the ISSSOMTransformApplication interface.
After a few trial-and-errors, I think _this_ is the final design for the
SSSOM/Transform parser (famous last words...).

Instead of making the parser abstract and asking applications to derive
it (which is slightly annoying because then the derived parser must
provide all the constructors again; also there's something wrong in
having a derived class that contains both the generic parsing logic and
the post-parsing application-specific logic), the parser is now concrete
again, and must be initialised with an implementation of the
ISSSOMTransformApplication interface, which provides all the
application-specific logic.