Fix processing of OBO 'treat-xrefs-as...' tags.
We recognise the 'treat-xrefs-as-genus-differentia' tag, which is
processed in exactly the same way as the '-reverse-genus-differentia'
onw -- the semapv:crossSpeciesExactMatch predicate does not distinguish
a mapping from a taxon-specific ontology to a taxon-neutral ontology and
a mapping in the other direction.

We use skos:exactMatch instead of owl:equivalentClass as the mapping
predicate associated with 'treat-xrefs-as-equivalent', for more
flexibility (the OWL translation can then still be changed in a SSSOM/T
rule).

Likewise, we associate skos:narrowMatch to 'treat-xrefs-as-has-subclass'
and skos:broadMatch to 'treat-xrefs-as-is_a'.