Allow to use mapping-dependent variables in SSSOM/T-Owl.
Generating the bridges for Uberon/CL presents an unforeseen difficulty
in that the relation to use for equivalence axioms may depend on whether
the object of the mapping is a subclass of one of Uberon's life stage
terms (for those mapping the equivalence axiom should use a BFO:0000066
relation instead of BFO:0000050).

To deal with that, we introduce the concept (in SSSOM/T-Owl, _not_ in
the generic, application-neutral SSSOM/T) of "mapping-dependent
variables". Such a variable has a default value (to be used in most
cases) and may also have subject-specific values and object-specific
values. When the variable is expanded, its value will depend on the
mapping that is currently being processed.

(This will probably require much more documentation than that to be
usable; then again, I expect most use cases of SSSOM/T-Owl will never
need such a feature.)