Amend the mapping-dependent variable feature.
Use a more generic design based on re-using the IMappingFilter
interface. A mapping-dependent variable is now a variable that has a set
of filters, each filter being associated with a value. To get the value
of the variable for a given mapping, apply each filter in turn on the
mapping until one returns true -- the value of the variable is the value
associated with that filter.

In SSSOM/T-Owl, use a simpler syntax that requires only one header
function "set_var", which can take either 2 arguments (the variable name
and its default value) or 3 (the variable name, its value, and the
condition associated with the value).