Allow to use the "fast path" with variable expressions.
When processing the "common pattern" expression for an equivalence
axiom, check whether either the relation IRI or the filler IRI are
registered variables, and expand them appropriately at run time.

This means we can now use a variable in a create_axiom call as follows:

  create_axiom('%subject_id EquivalentTo: %object_id and
               (<%VARIABLE> some NCBITaxon:7227)');

and the expression will still be recognised and be allowed to go through
the "fast path" (EquivalentAxiomGenerator) instead of having to go
through the Manchester parser.