Allow the AxiomDispatchTable to resolve entities declared in SSSOM/T.
Make the SSSOMTOwl class expose its custom entity checker so that we can
pass it to the AxiomDispatchTable parsing method. This is so that we can
resolve any class or object property that are not present in the
ontology but which have been declared in the SSSOM/T header with
'declare_class' or 'declare_object_property' functions.

A probably cleaner design would be take the CustomEntityChecker entirely
out of SSSOMTOwl: let the client code create it, initialise it with the
identifiers from an ontology, then pass it to both the SSSOMTOwl and the
AxiomDispatchTable. But that would be for a later refactoring.

Of note, the --dispatch-table option in sssom-inject is now only
available when --ruleset is also used. This is fine since this option is
only meaningful with a ruleset anyway (the dispatch table works with
tagged SSSOM/T rules, and the only way to get tagged rules in
sssom-inject is to get them from a ruleset file).