Add --extract option to sssom-inject.
Add an option '--extract' to sssom-inject to obtain mappings directly
from the cross-references in the input ontology. This is equivalent to
chaining the xref-extract and sssom-inject commands, but without having
to write the mappings to a file in the first command just to read them
back from that file in the second command.

This does not allow to control how mappings are extracted, though -- it
only mimics the default behaviour of the xref-extract command.