Add preprocessing functions to check existence.
Add a check_subject_existence() preprocessing function to SSSOM/T-Owl,
which returns the current mapping (allowing to be processed further) if
the subject exists in the ontology and is not obsolete, or null
(removing the mapping from the set) otherwise.

Likewise, add a check_object_existence() do check whether the object of
a mapping exists and is not obsolete.

Such functions may be necessary in cases where we need to invert the
mappings as part of the processing, and we can only check for (say) the
subject's existence after having inverted the mappings -- because in
that case we cannot simply use sssom-inject --check-subject option,
which would drop mappings to inexistent/obsolete subjects before the
inverting rule could be applied.