Only use 'add-axiom'-provided axioms if needed.
As currently implemented, if an entry in the axiom dispatch table
contains one or several 'add-axiom' field(s), then that entry will
systematically yield an ontology even if no axiom has been dispatched to
it, because the axiom(s) set by 'add-axiom' are taken into account when
deciding whether to create a file for an entry.

This commit changes this behaviour by storing the axioms set by
'add-axiom' into a separate set. When the time comes to save the
contents of the dispatch table, only entries that actually received
axioms through the addAxiom() method will get saved -- at that time,
any additional axiom that may have been set by 'add-axiom' will be saved
as well.