Add the '--bridge-iri' option.
Add a new option to allow specifying the ontology IRI of the bridge file
to be created with the '--bridge-file' option.

If the ontology IRI is already known to the ontology manager (which can
happen for example if the bridge file to be (re)generated is an import
of the current ontology), the corresponding ontology is emptied of its
current axioms which are then replaced by the axioms generated by the
command.