Add options to deal with duplicated cross-references.
It may happen that an ontology contain more than one cross-reference
pointing to the same foreign term. This may cause some problem for
downstream applications, so we add options to allow avoiding generating
mappings in such cases:

* --drop-duplicates: will ignore all cross-references pointing to the
  same foreign term, if there are more than one;
* --first-only: will generate a mapping only for the first
  cross-reference pointing to a foreign term (any subsequent
  cross-reference to that same term will be ignored).

(The wording used, "duplicates", is not great: those cross-references
are not really exact duplicates, only the object is duplicated. But it's
2.34am, so that will do for now.)