Log more useful messages for dropped duplicates.
Under --drop-duplicates, we get from the XrefExtractor the full list of
foreign terms that are mapped to more than one subject term. We can use
it to print more useful warning messages, where we give on a single line
for each object term the list of all the mapped subject terms.