More options to deal with duplicate mappings.
Allow sssom-inject to also deal with "duplicate" mappings, with two
options (--drop-multiple-subjects and --drop-multiple-objects) to drop
any mapping that has the same subject ID (resp. object ID) as any
previously processed mapping.

Also add the drop_multiple_subjects() and drop_multiple_objects()
preprocessing functions to do the same as part of the SSSOM/T-Owl
ruleset.

Note that because of the design of the SSSOM/T processing engine, it is
not possible to drop *all* mappings that have the same subject or
object. To do so would require to be able to "look forward" in the set
when processing a mapping to check whether other mappings have the same
subject or object, or to do a two-pass processing. Not worth the extra
complexity.