Allow to filter on mapping cardinality.
Update the SSSOM/T parser to recognise a 'mapping_cardinality==' filter.
When inverting a mapping, if it has a cardinality value make sure to
invert it as weell (so that inverting a mapping with, say, a '1:n'
cardinality yields an inverted mapping with a cardinality of 'n:1').

In sssom-inject, infer the cardinality of all mappings in the set prior
to any processing, so that we are ready to filter on cardinality.