Remove xref-extract' --first-only option.
Update the implementation of the xref-extract command to use
automatically inferred mapping cardinality to filter out "duplicate"
mappings, instead of ad-hoc code in the xref extractor. As a
side-effect, we remove the possibility of keeping only the first
"duplicate" mapping: we either remove all of them or none at all.