Write mappings (almost) deterministically.
Sort the mappings before writing them to a TSV file. We sort on subject,
predicate. and object, and then on a handful of other metadata fields.
This should be enough to ensure deterministic sorting for most mapping
sets. We are *not* going to sort on every possible fields -- not until I
see mapping sets in the wild containing some mappings that can't be
distinguished using the fields we do sort on.