Fix the mapping cardinality logic (again).
The computation of mapping cardinality was still not completely correct.

To get the subject side of the cardinality (X in X:Y), we need to figure
out how many *different* subjects there are for a same object, instead
of just how many times the same object appears in all mappings.
Likewise, to get the object side (Y in X:Y), we need to know how many
*different* objects there are for a same subject.

Otherwise, two mappings with the same subject and the same subject would
both be considered to be n:n mappings, while they really are 1:1
mappings -- cardinality is only about subject and object, independently
of all mapping metadata such as predicate, justification, etc.