Add --replace and --append options to xref-extract.
Add two options to xref-extract: --replace and --append. They both
change the way the --mapping-file option is interpreted: when any of
those two options is used, the file indicated by --mapping-file is
required to already exist and contain a valid mapping set, which the
command will read first before attempting to extract mappings from the
ontology.

With --append, the mappings extracted from the ontology will be appended
to the mappings already present in the file.

With --replace, all mappings already present in the file will be
discarded (actually, they won't be read at all) and replaced by the
mappings extracted from the ontology. This option allows to reuse the
metadata from an existing set, without carrying over the mappings
themselves.