Add SlotVisitor and associated classes.
This commit is an important overhaul of the way we deal with the SSSOM
"slots". Basically, we add a pseudo-visitor interface that can be used
to "visit" all slots in a MappingSet or Mapping object, without having
to bother about all the reflection code needed to enumerate those slots
and accessing their values.

(It is a "pseudo" visitor because, while it is intended to be used as a
classic Visitor pattern, it works very differently, since the objects we
are visiting are not classes but fields within classes.)

The new interface is used whenever possible in the rest of the code. In
particular, it makes the TSVReader much simpler. The TSVWriter is still
a bit more complicated than I would like, but now this is mostly due to
the three inner classes implementing the visitor interface, so it is
cleaner than the previous implementation. In the DirectAxiomGenerator
class, it allows to generate annotations for *all* the metadata slot,
with again all the complexity residing in the visitor implementation.