Sort Curie map entries.
When writing the Curie map in the metadata block, sort the prefix names
lexicographically. The order does not really matter, as long as it is
predictable and not dependent on the underlying hash map implementation.

This is neither required nor recommended by the SSSOM spec, but since
the spec does recommend sorting the slots in the metadata block (to
minimise the risk of spurious diffs), it makes sense to *also* sort the
contents of the prefix map.