Do not deserialise empty values to an empty string.
When reading a mapping set, if a column does not have an explicit value
for a given mapping, set the corresponding slot in that mapping to null
rather than to an empty string. Non-string slots (e.g.
"mapping_confidence" or "predicate_modifier") are already set to null
when there is no value, so string slots should be treated similarly.