Write metadata slots in spec order.
The current SSSOM spec does not actually say anything (yet) about the
order in which slots should appear in the metadata block, but Nico has a
stated preference for "spec order" (the order in which the slots are
listed in the spec) over alphabetical order, so we switch to that order.

Note as "curie_map" is *not* part of the model, it is unclear where it
should appear in that order. Here, we assume it should appear first.
Having the Curie map before any other slot could be useful for
implementations that want to expand CURIEs in the metadata "on the fly"
(while the metadata are being read).