Fully implement sorting of mappings.
Re-write the DefaultMappingComparator. Instead of trying to manually
compare a selection of slots, we use the SlotVisitor interface to create
a string representation of each mapping, then we compare the string
representations directly.

This allows to take into accounts *all* slots, instead of just a
selection of them, and therefore fully implements the sorting order
recommended by the specification.