Write the effective Curie map.
When writing the YAML metadata block, write the *effective* Curie map,
containing only those prefixes that are actually used in the set. Any
prefix found in the map but not used anywhere will not be written.

Also add a setPrefixMap() method to the TSVWriter to allow client code
to specify the map to use without having to modify the map associated
with the set.