Remove enum-specific methods in the SlotVisitor interface.
The three methods used to visit slots that have the type of a SSSOM
enumeration are of little use. We remove them and replace them by a
single method that will be used to visit all enum-typed slots regardless
of the exact type of the enum.

This will require a bit more work on the part of implementers if they do
need to act differently depending on the enum type (they will need to
check the slot type to determine whether the enum is MappingCardinality,
EntityType, or PredicateModifier), but in most cases they shouldn't need
to do that (that's true for all the internal uses of the interface).