Update documentation.
Add an example of how to use the SlotHelper class in its Javadoc
comment. Explain that the SSSOMTransformerReader.getRules() method
always returns all the rules parsed since the beginning of the lifetime
of the object. Update the website documentation in various places,
notably to: mention the SlotHelper class and the SlotVisitor interface,
mention the cardinality inference feature, and explain what a SSSOM/T
dialect is.