Implement missing filters in SSSOM/T.
Complete the SSSOM/T filtering feature by adding the possibility to
filter on previously non-filterable slots. This includes all the slots
that accept a string or an URI (rather than an identifier, a numerical
value, or an enum value), as well as the two slots that accept an
EntityType value.