Add new SimpleSlotVisitor interface.
Adds a new SimpleSlotVisitor interface that is, as the name implies, a
simplified version of SlotVisitor. It defines only one method to visit
all slots regardless of the slot types. In situations where the client
code does not need to get typed slot values, this interface offers a
less cumbersome way to visit slots, especially as it can be implemented
as a lambda method.