Propagate/condense slots when reading/writing a TSV file.
When reading a set from a TSV file, automatically propagate any slot
that can be propagated from the set level to the individual mappings.
This is done using the NeverReplace policy (only propagate a slot if it
is unset for *all* mappings, otherwise ignore). The policy may change in
the future depending on what is decided by the authors of the SSSOM
specification (but for now, NeverReplace looks likely to be the chosen
policy).

Conversely, when writing a set to a TSV file, automatically condense any
slot that can be condensed (if all mappings have the same value for that
slot).