[cli] Allow applying a SSSOM/T ruleset.
Add a --ruleset option to the CLI tool to apply a SSSOM/T ruleset to the
mappings before writing the result set.

The SSSOM/T application used by this option is a very simple one that
has only one "generating" function, "include", which instructs the
processor to include the current mapping into the result set.