Use dc/terms rather than dc/elements/1.1 for dispatch table metadata.
Update sssom-inject so that dc-title, dc-description, etc. fields in a
dispatch table are translated into annotations in the purl.org/dc/terms/
namespace, rather than the corresponding annotations in the
purl.org/dc/elements/1.1/ namespace.

Ideally this should somehow be configurable at runtime, but this entire
code is almost exclusively motivated by Uberon and may very well never
be used by anybody else, so we'll hardcode the annotations for now. The
Dublin Core Metadata Initiative is "gently encouraging the use of the
/terms/ namespace" anyway, so if anyone complains about the hardcoding
we can just say that we are gently following the DCMI recommendations.