Do not map directly to a SSSOM object when parsing.
When parsing the metadata block of a mapping set, do not map the YAML
object directly to a MappingSet object. Map it to a generic
Map<String,Object>, then iterate over the contents to manually fill the
fields of the MappingSet object.

Likewise, when parsing the TSV rows, do not map each row directly to a
Mapping object. Use a generic Map<String,Object> intermediate.

This is more cumbersome than straight mapping, but it allows custom
treatment of some fields if needed and to ignore unknown fields, thereby
making the parser ready to deal with possible evolutions of the SSSOM
format and/or data model.