Add unit test for parsing of pseudo-list values.
Test that we can successfully parse a file that misuses multi-valued
slots as if they were single-valued.

That is, we must parse things such as:

creator_id: "value"

as the equivalent of:

creator_id:
  - "value"

This is a common error and even the SSSOM documentation gets it wrong,
so we should be liberal when encountering it.