Rename the sssom-inject command to "inject" only.
The recommended way to use the ROBOT commands provided by this project
are through the use of a plugin (the standalone version is mostly
intended for development and debugging). When used as a plugin, the
command name will already be prefixed by the plugin's basename
(typically 'sssom'), so having 'sssom-' in the command name is
redundant.

In the standalone version, we keep the old 'sssom-inject' name though,
as 'inject' on its own is not explicit enough.