Invert the logic of the mapping processor.
Instead of iterating over mappings first and then iterating over the
processing rules for each mapping, we iterate over the processing rules
first and then iterate over the mappings for each rule.

This is necessary because if a rule leads to some mappings being
dropped, we may need to re-calculate the mapping cardinality of all the
remaining mappings before another rule that filters on cardinality can
be applied correctly.