Infer license and ID from the ontology when extracting xrefs.
When extracting a mapping set from cross-references in an ontology, if
no license is specified either on the command line or in a pre-existing
metadata table, try using the license of the ontology itself.

Likewise, if no set ID is specified either on the command line or in a
pre-existing metadata table, automatically construct an ID from the IRI
of the ontology.