[cli] Allow speciying SSSOM/T rules from the command line.
Add a -R, --rule option to sssom-cli to specify a single SSSOM/T rule on
the command line. This is to allow avoiding the use of a SSSOM/T file
for simple filtering tasks.

The option may be used several times to specify several rules. It may
also be combined with the --ruleset option to load rules both from a
file and from the command line -- in that case, the file is read first,
which means that the rules from the file will be applied before any rule
read from the command line.