[cli] Add an option to include all mappings by default.
Add a -a, --include-all option to sssom-cli. When enabled, this option
automatically adds a rule at the end of the effective ruleset to include
all mappings that have not been previously dropped. This basically
changes the behaviour of the sssom-cli filter from "only keep explicitly
included mappings" to "only drop explicitly excluded mappings".