Fix null pointer exception when writing the Curie map.
The list of "used prefixes" may contain a null value, if an IRI could
not be shortened by the prefix manager because no matching prefix is
known. When that happens, we must take care of explicitly excluding the
null prefix to avoid a null pointer exception when we sort the entries.