Allow mixed uses of addEdit and transform.
When `addEdit()` is called,  make sure to reset the slot helper used to
iterate over the slots to be edited. This allows to use the same editor
object to make new changes even after its `transform()` method has been
called once -- this may not happen in the context of SSSOM/T but could
happen elsewhere.