Allow to customise axiom annotations in "direct" mode.
When using the "direct" function in SSSOM/T-OWL, allow an optional
argument to describe how to annotate the generated axiom using the
mapping metadata. This optional argument works in exactly the same way
as the similar argument to all the other axiom-generating function.

The default behaviour of the "direct" function, when no optional
argument is provided, is still to generate "direct" annotations for all
metadata except "mapping_cardinality". To completely disable axiom
annotations (i.e. to only generate an unnannotated axiom representing
the mapping itself), it is possible to provide an empty list of metadata
slots, as in "direct:".