Add the possibility of performing regex-based replacement on mappings.
This commit adds a new preprocessing function called "replace". It
accepts three arguments: the name of a metadata slot, a pattern to find
in the value of that slot, and a string to replace the pattern with.

The function is implemented using the same MappingEditor object used for
the "edit" function. The MappingEditor has been slightly modified so it
can be used to apply "delayed" changes -- changes whose values are only
known at the time the transformer is applied to a mapping.