Allow to edit more than one slots with assign/replace.
Modify the "assign" and "replace" functions so that they can be used to
edit more than one slot in a single call, by accepting an unlimited
number of arguments.

Update the documentation with the two new functions. Remove any mention
of the "edit()" function, which is still supported but deprecated in
favour of the "assign()" function.