cli: Allow using the original Curie map.
When writing the resulting set, provide the writer with a prefix map
that contains both the prefixes from the original set (the *first* set,
if several sets were read) plus any prefixes from the mapping processor
(those found in a "prefix" declaration in the SSSOM/T header and those
given on the command line). The prefixes from the mapping processor take
precedence, so the user can always override any prefix declared in the
original set.