Add --error-on-unshortenable-iris to the inject command.
When generating annotation axioms using %subject_curie or %object_curie,
record any IRI that cannot be shortened, and print a warning for each of
them once processing is over.

If the command is called with --error-on-unshortenable-iris, then we
error out after printing the corresponding warnings, and before writing
any result file.