cli: Do not include cardinality slot by default.
When writing the result set, forcibly set all "mapping_cardinality"
slots to null so that they don't get written in the output. If we have
merged several sets and/or performed any kind of filtering, the original
values may no longer be accurate, so it's better not to write them at
all.

A new option "--force-cardinality" is added to force cardinality values
to be written. In that case, cardinality is re-calculated just before
writing the set, to ensure the values are correct.