cli: Add the --split option.
Add an option to sssom-cli to split the result set along the prefix name
of the subjects and objects. That is, the option creates one file for
each couple {subject prefix name, object prefix name}. When that option
is used, the normal output (where the entire result set is written to
standard output or to the file indicated by the --output option) is
disabled.