cli: Allow to read metadata from a specific file.
If the argument to the --input option contains a ':' character, we split
the argument in two on that character. The left part is interpreted as
the name of the TSV file, and the right part as the name of the metadata
file. This allows to read a set that has its metadata in a file that
does not follow the recommended naming convention.