cli: Merge the set-level metadata from all the input sets.
When reading several input sets, merge the set-level metadata from all
the input sets.  Only multi-valued slots are merged: for the other
slots, only the values from the first set are used for the output set.

Use the --no-metadata-merge option to disable that behaviour and only
use the metadata from the first set, multi-valued or not.