Write declared "extra" slots.
This commit implements the writing part of the proposed behaviour in
mapping-commons/sssom#328 for dealing with non-standard slots.

If the __extra special dictionary at the mapping set level contains
entries whose keys are declared in extra_set_slots, those entries are
written with the set metadata as if they were normal slot. They are
written in key alphabetical order, after all the other slots.

If the __extra special dictionary at the mapping level contains entries
whose keys are declared in the extra_slots, those entries are written
with the mappings metadata in the TSV section as supplementary columns.
The columns are written in key alphabetical order, after all the other
columns.