Remove the extra_set_slots declaration slot.
Remove the need to declare extra slots for the set level. Instead, extra
set metadata should go into a "extra_metadata" slot, which is expected
to be a dictionary of strings.

Also, the extra_slots field is renamed to "extra_columns", to better
reflect what it does: declare that there are some extra, non-standard
columns in the TSV section.