Escape values in the TSV section.
When writing a TSV file, apply CSV-style (RFC 4180) quoting rules to
deal with the possible presence of tab or new line characters in TSV
values.

Such an escaping is not mandated by the spec (yet?), but is already
supported by sssom-py (through Pandas) and by our own parser (through
Jackson).