Add write support for extension slots.
This commit completes the implementation of the latest proposal for the
"non-standard metadata support" feature. It replaces the ExtensionHelper
class by an ExtensionSlotManager, extends the SlotVisitor interface and
the SlotHelper object to allow "visiting" the slots containing the
extension definitions and the extension themselves, and updates the TSV
writer to allow writing non-standard metadata slots.