Write TSV headers when writing an empty set.
When writing a set that happens to contain no mappings at all, make sure
to still write a default set of headers (just for the columns that are
mandatory: subject_id, predicate_id, object_id, and
mapping_justification).

This ensures that a set written by the TSVWriter can always be read back
correctly by the TSVReader, because the TSVReader cannot read a set in
the absence of headers.