Add a short name variant of the --prefix-map-from-input option.
I have found myself using the `--prefix-map-from-input` option more
often that I would have thought, so it is actually useful to have a
one-letter form for it.

For the same reason, we remove the "(not recommended") warning in the
help message of that option. Using the input prefix map should
definitely not be the default behaviour, but it does not have to be
actively discouraged.