Write non-standard metadata by default.
Change the default value of the `--write-extra-metadata` option to
DEFINED, so that if the mapping set to write contains non-standard
metadata, they would be written by default (as defined extension slots)
rather than discarded.

It is safe to do so since the default value of `--accept-extra-metadata`
is still NONE, meaning that non-standard metadata from the input set
would still be discarded by default unless another value is explicitly
specified.