cli: Add convenience --include/--exclude options.
Add two more options to filter mappings in or out without having to
specify the '-> include()' or '-> stop()' part of the rule.

The '--exclude' option accepts a SSSOM/T filter expression and simply
appends a '-> stop()' to it and then adds it to the overall ruleset.

Likewise, the '--include' option accepts a SSSOM/T filter expression and
appends a '-> include()' to it before adding to the overall ruleset.