Allow updating labels in a mapping set from an ontology.
Add an new option to sssom-cli, --update-from-ontology, to update the
labels of subjects and objects in a mapping set with labels found in a
specified ontology. The option may be specified many times, to use
several ontologies as source of information to update the labels.

The subject_source field of a mapping is also set to the ontology IRI
if: 1) the subject_label has been updated (meaning the ontology does
contain the subject entity in its signature) and 2) the subject_source
field was not set. Likewise for the object_source field.

close #6