cli: Allow removal of mappings to missing entities.
When updating a mapping set from an ontology, optionally allows removing
any mapping whose subject and/or object does not exist in the ontology,
or is marked as obsolete.

This is done by adding comma-separated flags after the filename of the
ontology ("--update-from-ontology file.owl:flag1,flag2"), with the
"subject" flag instructing the program to check for existence of the
mapping's subject, and the "object" flag instructinf to check for
existence of the mapping's object.

The method that performs the update in the core module allows for finer
selection of what is to be updated (e.g. only updating the labels, only
updating the source fields, only deleting for missing subject, only
deleting for obsolete subject, etc.), but at least for now all those
options are not available from the CLI.