cli: Add support for XML catalogs.
When loading an ontology to use for updating a mapping set, allow the
use of a XML catalog to resolve any import declarations.

The default catalog is "catalog-v001.xml"; if a file with such a name
exists in the current directory, it is automatically used. Another
catalog can be explicitly specified with the --catalog option.